requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.4.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
